new_scrapyproject/requirements.txt
Scrapy==2.8.0
psycopg2==2.9.2
jinja2==3.0.3
pyOpenSSL==21.0.0
Flask==2.0.2
python-dotenv==0.19.1
scrapyd-client==1.2.3
requests==2.26.0
psycopg2-binary==2.9.3
cryptography>=3.4.6
